+++ storybook/src/example-pages/componenten/application-heading/application-heading.ts
@@ -8,7 +8,7 @@ examplePageFactory(
   ({ applicationHeadingTemplate }) => html`
     <div class="container">
       ${applicationHeadingTemplate({
-        title: "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor.",
+        title: "Vergunningcheck",
       })}
     </div>
   `